Cleanup loader.cc
Several relics of debugging loader.cc were cleaned up. In particular,
the internal API of the read_file function used was switched to
return the size of the file instead of the boolean success value
(instead panicking if a requested file does not exist). This greatly
simplifies the logic of the loader.