Further progress on USB
Several bugs were fixed, several new registers are written too, and I
did a bit of research into the elusive LAN_RUN pin which seems to be
related -- the GPIO pin # differs between boards, but it's connected to
nRESET on the LAN9215 chip. Cycling the pin low-high should reset the
controller, in theory.